Clean up `RejectionInFilterTest`
