+++ server/src/test/java/io/spine/server/commandbus/given/caffetteria/CaffetteriaStats.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2020, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.commandbus.given.caffetteria;
+
+import io.spine.core.Subscribe;
+import io.spine.server.projection.Projection;
+import io.spine.test.commandbus.CmdBusCaffetteriaId;
+import io.spine.test.commandbus.CmdBusCaffetteriaStats;
+import io.spine.test.commandbus.command.CaffetteriaRejections;
+import io.spine.test.commandbus.event.CmdBusTableAllocated;
+
+public final class CaffetteriaStats extends Projection<CmdBusCaffetteriaId,
+                                                       CmdBusCaffetteriaStats,
+                                                       CmdBusCaffetteriaStats.Builder> {
+
+    @Subscribe
+    void on(CmdBusTableAllocated event) {
+        builder().setVisitorCount(state().getVisitorCount() + event.getVisitorCount());
+    }
+
+    @Subscribe
+    void on(CaffetteriaRejections.CmdBusEntryDenied rejection) {
+        builder().setEntryDenied(state().getEntryDenied() + rejection.getVisitorCount());
+    }
+}
